refactor: restructure the LLM components
