+++ src/solace_ai_connector/common/utils.py
@@ -125,9 +125,9 @@ def import_module(module, base_path=None, component_package=None):
                     ".components",
                     ".components.general",
                     ".components.general.for_testing",
-                    ".components.general.langchain",
-                    ".components.general.openai",
-                    ".components.general.litellm",
+                    ".components.general.llm.langchain",
+                    ".components.general.llm.openai",
+                    ".components.general.llm.litellm",
                     ".components.general.websearch",
                     ".components.inputs_outputs",
                     ".transforms",
