+++ src/solace_ai_connector/components/__init__.py
@@ -25,7 +25,7 @@ from .general.for_testing import (
     give_ack_output,
 )
 
-from .general.langchain import (
+from .general.llm.langchain import (
     langchain_embeddings,
     langchain_vector_store_delete,
     langchain_chat_model,
@@ -34,7 +34,7 @@ from .general.langchain import (
     langchain_vector_store_embedding_search,
 )
 
-from .general.litellm import (
+from .general.llm.litellm import (
     litellm_chat_model,
     litellm_chat_model_with_history,
 )
@@ -62,17 +62,17 @@ from .general.delay import Delay
 from .general.iterate import Iterate
 from .general.message_filter import MessageFilter
 from .general.parser import Parser
-from .general.langchain.langchain_base import LangChainBase
-from .general.langchain.langchain_embeddings import LangChainEmbeddings
-from .general.langchain.langchain_vector_store_delete import LangChainVectorStoreDelete
-from .general.langchain.langchain_chat_model import LangChainChatModel
-from .general.langchain.langchain_chat_model_with_history import (
+from .general.llm.langchain.langchain_base import LangChainBase
+from .general.llm.langchain.langchain_embeddings import LangChainEmbeddings
+from .general.llm.langchain.langchain_vector_store_delete import LangChainVectorStoreDelete
+from .general.llm.langchain.langchain_chat_model import LangChainChatModel
+from .general.llm.langchain.langchain_chat_model_with_history import (
     LangChainChatModelWithHistory,
 )
-from .general.langchain.langchain_vector_store_embedding_index import (
+from .general.llm.langchain.langchain_vector_store_embedding_index import (
     LangChainVectorStoreEmbeddingsIndex,
 )
-from .general.langchain.langchain_vector_store_embedding_search import (
+from .general.llm.langchain.langchain_vector_store_embedding_search import (
     LangChainVectorStoreEmbeddingsSearch,
 )
 from .general.websearch.websearch_duckduckgo import WebSearchDuckDuckGo
