+++ src/solace_ai_connector/components/general/llm/langchain/langchain_chat_model.py
