+++ src/solace_ai_connector/components/general/llm/langchain/langchain_chat_model_with_history.py
@@ -16,7 +16,7 @@ from langchain.schema.messages import (
     SystemMessage,
 )
 
-from ....common.message import Message
+from .....common.message import Message
 from .langchain_chat_model_base import (
     LangChainChatModelBase,
     info_base,
