+++ src/solace_ai_connector/components/general/llm/langchain/langchain_vector_store_embedding_index.py
