+++ src/solace_ai_connector/components/general/llm/litellm/litellm_chat_model.py
