+++ src/solace_ai_connector/components/general/llm/openai/openai_chat_model.py
